+++ src/adapter/graphql/src/queries/flows/flow_description.rs
@@ -18,8 +18,7 @@ use kamu_core::{
     MetadataQueryService,
     PollingSourceBlockInfo,
 };
-use kamu_flow_system as fs;
-use kamu_flow_system::FlowResultDatasetUpdate;
+use kamu_flow_system::{self as fs, FlowResultDatasetUpdate};
 
 use crate::prelude::*;
 
@@ -64,6 +63,7 @@ pub(crate) struct FlowDescriptionDatasetPushIngest {
     source_name: Option<String>,
     input_records_count: u64,
     ingest_result: Option<FlowDescriptionUpdateResult>,
+    message: String,
 }
 
 #[derive(SimpleObject)]
@@ -338,11 +338,16 @@ impl FlowDescriptionBuilder {
                     })
                 } else {
                     let source_name = flow_state.primary_trigger().push_source_name();
+                    let trigger_description = flow_state
+                        .primary_trigger()
+                        .trigger_source_description()
+                        .unwrap();
                     FlowDescriptionDataset::PushIngest(FlowDescriptionDatasetPushIngest {
                         dataset_id: dataset_key.dataset_id.clone().into(),
                         source_name,
                         input_records_count: 0, // TODO
                         ingest_result,
+                        message: trigger_description,
                     })
                 }
             }