+++ src/app/cli/src/app.rs
@@ -538,6 +538,11 @@ pub fn configure_base_catalog(
         kamu_datasets::MESSAGE_PRODUCER_KAMU_DATASET_DEPENDENCY_GRAPH_SERVICE,
     );
 
+    register_message_dispatcher::<kamu_datasets::DatasetExternallyChangedMessage>(
+        &mut b,
+        kamu_datasets::MESSAGE_PRODUCER_KAMU_HTTP_ADAPTER,
+    );
+
     b
 }