+++ src/domain/flow-system/domain/src/entities/flow/flow_trigger_type.rs
@@ -48,6 +48,32 @@ impl FlowTriggerType {
         }
     }
 
+    pub fn trigger_source_description(&self) -> Option<String> {
+        match self {
+            Self::Manual(_) => Some("Flow triggered manually".to_string()),
+            Self::AutoPolling(_) => Some("Flow triggered automatically".to_string()),
+            Self::Push(trigger_push) => match &trigger_push.result {
+                DatasetPushResult::HttpIngest(_) => {
+                    Some("Flow triggered by root dataset ingest via http endpoint".to_string())
+                }
+                DatasetPushResult::SmtpSync(sync_result) => {
+                    if let Some(account_name) = sync_result.account_name_maybe.as_ref() {
+                        Some(format!(
+                            "Flow triggered by root dataset ingest via SMTP sync by account: \
+                             {account_name}",
+                        ))
+                    } else {
+                        Some(
+                            "Flow triggered by root dataset ingest via SMTP sync anonymously"
+                                .to_string(),
+                        )
+                    }
+                }
+            },
+            Self::InputDatasetFlow(_) => Some("Flow triggered by completed root flow".to_string()),
+        }
+    }
+
     /// Checks if new trigger is unique compared to the existing triggers
     pub fn is_unique_vs(&self, existing_triggers: &[FlowTriggerType]) -> bool {
         // Try finding a similar existing trigger and abort early, when found
@@ -99,9 +125,30 @@ pub struct FlowTriggerAutoPolling {
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct FlowTriggerPush {
-    // TODO: source (HTTP, MQTT, CMD, ...)
     pub trigger_time: DateTime<Utc>,
     pub source_name: Option<String>,
+    pub dataset_id: odf::DatasetID,
+    pub result: DatasetPushResult,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub enum DatasetPushResult {
+    HttpIngest(DatasetPushHttpIngestResult),
+    SmtpSync(DatasetPushSmtpSyncResult),
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct DatasetPushHttpIngestResult {
+    pub old_head_maybe: Option<odf::Multihash>,
+    pub new_head: odf::Multihash,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct DatasetPushSmtpSyncResult {
+    pub old_head_maybe: Option<odf::Multihash>,
+    pub new_head: odf::Multihash,
+    pub account_name_maybe: Option<odf::AccountName>,
+    pub is_force: bool,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -142,7 +189,12 @@ mod tests {
         });
         static ref PUSH_SOURCE_TRIGGER: FlowTriggerType = FlowTriggerType::Push(FlowTriggerPush {
             trigger_time: Utc::now(),
-            source_name: None
+            source_name: None,
+            dataset_id: TEST_DATASET_ID.clone(),
+            result: DatasetPushResult::HttpIngest(DatasetPushHttpIngestResult {
+                old_head_maybe: None,
+                new_head: odf::Multihash::from_digest_sha3_256(b"some-slice")
+            }),
         });
         static ref INPUT_DATASET_TRIGGER: FlowTriggerType =
             FlowTriggerType::InputDatasetFlow(FlowTriggerInputDatasetFlow {
@@ -203,7 +255,12 @@ mod tests {
         assert!(
             PUSH_SOURCE_TRIGGER.is_unique_vs(&[FlowTriggerType::Push(FlowTriggerPush {
                 trigger_time: Utc::now(),
-                source_name: Some("different".to_string())
+                source_name: Some("some-source".to_string()),
+                dataset_id: TEST_DATASET_ID.clone(),
+                result: DatasetPushResult::HttpIngest(DatasetPushHttpIngestResult {
+                    old_head_maybe: None,
+                    new_head: odf::Multihash::from_digest_sha3_256(b"some-slice")
+                }),
             })])
         );