+++ src/e2e/app/cli/common/src/kamu_api_server_client_ext.rs
@@ -1455,7 +1455,14 @@ impl FlowApi<'_> {
         }
     }
 
-    pub async fn wait(&self, dataset_id: &odf::DatasetID) {
+    // Method to wait for a flow to finish
+    // Args:
+    // - dataset_id: The ID of the dataset to check
+    // - expected_flow_count: The exact number of flows to wait for. Useful for
+    //   testcases
+    // where we have 2 or more flows for specific dataset which are triggered
+    // asynchronously For most cases will be used 1
+    pub async fn wait(&self, dataset_id: &odf::DatasetID, expected_flow_count: usize) {
         let retry_strategy = FixedInterval::from_millis(5_000).take(18); // 1m 30s
 
         Retry::spawn(retry_strategy, || async {
@@ -1492,6 +1499,9 @@ impl FlowApi<'_> {
             let edges = response["datasets"]["byId"]["flows"]["runs"]["listFlows"]["edges"]
                 .as_array()
                 .unwrap();
+            if edges.len() < expected_flow_count {
+                return Err(());
+            }
             let all_finished = edges.iter().all(|edge| {
                 let status = edge["node"]["status"].as_str().unwrap();